requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via
    #   -r requirements.in
    #   flask
boto3==1.36.2
    # via -r requirements.in
botocore==1.36.2
    # via
    #   boto3
    #   s3transfer
certifi==2024.12.14
    # via
    #   -r requirements.in
    #   requests
cffi==1.17.1
    # via
    #   -r requirements.in
    #   cryptography
charset-normalizer==3.4.1
    # via
    #   -r requirements.in
    #   requests
click==8.1.8
    # via
    #   -r requirements.in
    #   flask
colorama==0.4.6
    # via
    #   -r requirements.in
    #   click
cryptography==44.0.0
    # via
    #   -r requirements.in
    #   pyopenssl
flask==3.1.0
    # via -r requirements.in
fonttools==4.55.4
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via
    #   -r requirements.in
    #   requests
itsdangerous==2.2.0
    # via
    #   -r requirements.in
    #   flask
jinja2==3.1.5
    # via
    #   -r requirements.in
    #   flask
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
    #   werkzeug
packaging==24.2
    # via gunicorn
pillow==11.1.0
    # via -r requirements.in
pycparser==2.22
    # via
    #   -r requirements.in
    #   cffi
pyopenssl==25.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
s3transfer==0.11.1
    # via boto3
six==1.17.0
    # via
    #   -r requirements.in
    #   python-dateutil
typing-extensions==4.12.2
    # via -r requirements.in
urllib3==2.3.0
    # via
    #   -r requirements.in
    #   botocore
    #   requests
werkzeug==3.1.3
    # via
    #   -r requirements.in
    #   flask
